+++ CHANGELOG.md
@@ -23,6 +23,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Fixed
 * Python2 issues with `Boxes.__repr__`
+* Can now correctly draw 1D `Boxes` objects
 * Python2 issues mask shape using List[long] instead of List[int]
 * Zero division in Cython `non-maximum-supression` with zero sized boxes.
 * `Heatmap.shape` now works even if `class_probs` is not set.